Separated server stuff from client stuff.
