+++ src/Neptunium/App.xaml.cs
@@ -8,6 +8,7 @@ using Neptunium.View;
 using Neptunium.View.Dialog;
 using Neptunium.ViewModel;
 using Neptunium.ViewModel.Dialog;
+using Neptunium.ViewModel.Server;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -233,8 +234,16 @@ namespace Neptunium
 
         public override async Task OnFreshLaunchAsync(LaunchActivatedEventArgs args)
         {
-            WindowManager.GetNavigationManagerForCurrentWindow()
-                .RootNavigationService.NavigateTo<AppShellViewModel>();
+            if (!NepApp.IsServerMode)
+            {
+                WindowManager.GetNavigationManagerForCurrentWindow()
+                    .RootNavigationService.NavigateTo<AppShellViewModel>();
+            }
+            else
+            {
+                WindowManager.GetNavigationManagerForCurrentWindow()
+                    .RootNavigationService.NavigateTo<ServerShellViewModel>();
+            }
 
             await PostUIInitAsync();
         }
@@ -243,8 +252,16 @@ namespace Neptunium
         {
             if (args.PreviousExecutionState != ApplicationExecutionState.Running)
             {
-                WindowManager.GetNavigationManagerForCurrentWindow()
-                     .RootNavigationService.SafeNavigateTo<AppShellViewModel>();
+                if (!NepApp.IsServerMode)
+                {
+                    WindowManager.GetNavigationManagerForCurrentWindow()
+                        .RootNavigationService.SafeNavigateTo<AppShellViewModel>();
+                }
+                else
+                {
+                    WindowManager.GetNavigationManagerForCurrentWindow()
+                        .RootNavigationService.SafeNavigateTo<ServerShellViewModel>();
+                }
             }
 
             if (args.Kind == ActivationKind.Protocol)
@@ -346,16 +363,19 @@ namespace Neptunium
 
         protected override Task OnSuspendingAsync()
         {
-            if (App.GetDevicePlatform() == Crystal3.Core.Platform.Desktop || App.GetDevicePlatform() == Crystal3.Core.Platform.Mobile)
+            if (!NepApp.IsServerMode)
             {
-                //clears the tile if we're suspending.
-                TileUpdateManager.CreateTileUpdaterForApplication().Clear();
-            }
+                if (App.GetDevicePlatform() == Crystal3.Core.Platform.Desktop || App.GetDevicePlatform() == Crystal3.Core.Platform.Mobile)
+                {
+                    //clears the tile if we're suspending.
+                    TileUpdateManager.CreateTileUpdaterForApplication().Clear();
+                }
 
-            if (!NepApp.MediaPlayer.IsPlaying)
-            {
-                //removes the now playing notification from the action center.
-                ToastNotificationManager.History.Remove(NepAppUIManagerNotifier.SongNotificationTag);
+                if (!NepApp.MediaPlayer.IsPlaying)
+                {
+                    //removes the now playing notification from the action center.
+                    ToastNotificationManager.History.Remove(NepAppUIManagerNotifier.SongNotificationTag);
+                }
             }
 
             return Task.CompletedTask;
