+++ src/Neptunium/View/ServerShellView.xaml
@@ -1,5 +1,5 @@
 ﻿<Page
-    x:Class="Neptunium.View.IoTShellView"
+    x:Class="Neptunium.View.ServerShellView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:Neptunium.View"
