+++ src/Neptunium/ViewModel/AppShellViewModel.cs
@@ -54,99 +54,49 @@ namespace Neptunium.ViewModel
 
         public AppShellViewModel()
         {
-            if (App.GetDevicePlatform() != Crystal3.Core.Platform.IoT)
-            {
-                NepApp.UI.AddNavigationRoute("Stations", typeof(StationsPageViewModel), ""); //"");
-                NepApp.UI.AddNavigationRoute("Now Playing", typeof(NowPlayingPageViewModel), "");
-                NepApp.UI.AddNavigationRoute("History", typeof(SongHistoryPageViewModel), "");
-                NepApp.UI.AddNavigationRoute("Schedule", typeof(StationProgramsPageViewModel), "");
-                NepApp.UI.AddNavigationRoute("Settings", typeof(SettingsPageViewModel), "");
+
+            NepApp.UI.AddNavigationRoute("Stations", typeof(StationsPageViewModel), ""); //"");
+            NepApp.UI.AddNavigationRoute("Now Playing", typeof(NowPlayingPageViewModel), "");
+            NepApp.UI.AddNavigationRoute("History", typeof(SongHistoryPageViewModel), "");
+            NepApp.UI.AddNavigationRoute("Schedule", typeof(StationProgramsPageViewModel), "");
+            NepApp.UI.AddNavigationRoute("Settings", typeof(SettingsPageViewModel), "");
 
 #if !DEBUG
                 if ((bool)NepApp.Settings.GetSetting(AppSettings.ShowRemoteMenu))
                 {
 #endif
-                    NepApp.UI.AddNavigationRoute("Remote", typeof(ServerRemotePageViewModel), "");
+            NepApp.UI.AddNavigationRoute("Remote", typeof(ServerRemotePageViewModel), "");
 #if !DEBUG
                 }
 #endif
 
-                NepApp.UI.AddNavigationRoute("About", typeof(AboutPageViewModel), "");
-            }
 
             NepApp.MediaPlayer.IsPlayingChanged += Media_IsPlayingChanged;
             NepApp.SongManager.PreSongChanged += SongManager_PreSongChanged;
             NepApp.SongManager.SongChanged += SongManager_SongChanged;
             NepApp.SongManager.StationRadioProgramStarted += SongManager_StationRadioProgramStarted;
 
-            if (App.GetDevicePlatform() != Crystal3.Core.Platform.IoT)
+            if (UserProfilePersonalizationSettings.IsSupported())
             {
-                if (UserProfilePersonalizationSettings.IsSupported())
-                {
-                    NepApp.SongManager.SongArtworkAvailable += SongManager_SongArtworkAvailable;
-                    NepApp.SongManager.NoSongArtworkAvailable += SongManager_NoSongArtworkAvailable;
-                }
-            }
-            else
-            {
-                //if its IoT
-                NepApp.ServerFrontEnd.DataReceived += ServerFrontEnd_DataReceived;
-            }
-        }
-
-        private async void ServerFrontEnd_DataReceived(object sender, NepAppServerFrontEndManager.NepAppServerFrontEndManagerDataReceivedEventArgs e)
-        {
-            if (!string.IsNullOrWhiteSpace(e.Data))
-            {
-                string data = e.Data.Trim();
-                string[] dataBits = data.Split(NepAppServerFrontEndManager.NepAppServerClient.MessageTypeSeperator);
-
-                if (dataBits.Length > 0)
-                {
-                    switch(dataBits[0].ToUpper())
-                    {
-                        case "PLAY":
-                            if (dataBits.Length > 1)
-                            {
-                                string stationName = dataBits[1];
-                                var station = await NepApp.Stations.GetStationByNameAsync(stationName);
-                                if (station != null)
-                                {
-                                    var stream = station.Streams.First();
-                                    await await App.Dispatcher.RunAsync(async () =>
-                                    {
-                                        await NepApp.MediaPlayer.TryStreamStationAsync(stream);
-                                    });
-                                    
-                                }
-                            }
-
-                            break;
-                        case "STOP":
-                            NepApp.MediaPlayer.Pause();
-                            break;
-                    }
-                }
+                NepApp.SongManager.SongArtworkAvailable += SongManager_SongArtworkAvailable;
+                NepApp.SongManager.NoSongArtworkAvailable += SongManager_NoSongArtworkAvailable;
             }
         }
 
         private async void SongManager_NoSongArtworkAvailable(object sender, Media.Songs.NepAppSongMetadataArtworkEventArgs e)
         {
-            if (App.GetDevicePlatform() != Crystal3.Core.Platform.IoT)
+            if ((bool)NepApp.Settings.GetSetting(AppSettings.UpdateLockScreenWithSongArt))
             {
-                if ((bool)NepApp.Settings.GetSetting(AppSettings.UpdateLockScreenWithSongArt))
+                //sets the fallback lockscreen image when we don't have any artwork available.
+                if (e.ArtworkType == Media.Songs.NepAppSongMetadataBackground.Artist)
                 {
-                    //sets the fallback lockscreen image when we don't have any artwork available.
-                    if (e.ArtworkType == Media.Songs.NepAppSongMetadataBackground.Artist)
+                    try
                     {
-                        try
-                        {
-                            await NepApp.UI.LockScreen.TrySetFallbackLockScreenImageAsync();
-                        }
-                        catch (Exception)
-                        {
+                        await NepApp.UI.LockScreen.TrySetFallbackLockScreenImageAsync();
+                    }
+                    catch (Exception)
+                    {
 
-                        }
                     }
                 }
             }
@@ -154,28 +104,25 @@ namespace Neptunium.ViewModel
 
         private async void SongManager_SongArtworkAvailable(object sender, Media.Songs.NepAppSongMetadataArtworkEventArgs e)
         {
-            if (App.GetDevicePlatform() != Crystal3.Core.Platform.IoT)
+            if ((bool)NepApp.Settings.GetSetting(AppSettings.UpdateLockScreenWithSongArt))
             {
-                if ((bool)NepApp.Settings.GetSetting(AppSettings.UpdateLockScreenWithSongArt))
+                if (e.ArtworkType == Media.Songs.NepAppSongMetadataBackground.Artist)
                 {
-                    if (e.ArtworkType == Media.Songs.NepAppSongMetadataBackground.Artist)
+                    try
                     {
-                        try
-                        {
-                            bool result = await NepApp.UI.LockScreen.TrySetLockScreenImageFromUriAsync(e.ArtworkUri);
+                        bool result = await NepApp.UI.LockScreen.TrySetLockScreenImageFromUriAsync(e.ArtworkUri);
 
-                            if (!result)
-                            {
-                                await NepApp.UI.LockScreen.TrySetFallbackLockScreenImageAsync();
-                            }
-                        }
-                        catch (Exception)
+                        if (!result)
                         {
-                            //todo make and set an image that represents the lack of artwork. maybe a dark image with the app logo?
-                            //maybe allow the user to set an image to use in this case.
-
                             await NepApp.UI.LockScreen.TrySetFallbackLockScreenImageAsync();
                         }
+                    }
+                    catch (Exception)
+                    {
+                        //todo make and set an image that represents the lack of artwork. maybe a dark image with the app logo?
+                        //maybe allow the user to set an image to use in this case.
+
+                        await NepApp.UI.LockScreen.TrySetFallbackLockScreenImageAsync();
                     }
                 }
             }
@@ -183,51 +130,45 @@ namespace Neptunium.ViewModel
 
         private async void SongManager_StationRadioProgramStarted(object sender, Media.Songs.NepAppStationProgramStartedEventArgs e)
         {
-            if (App.GetDevicePlatform() != Crystal3.Core.Platform.IoT)
+            if ((bool)NepApp.Settings.GetSetting(AppSettings.ShowSongNotifications))
             {
-                if ((bool)NepApp.Settings.GetSetting(AppSettings.ShowSongNotifications))
+                if (!await App.GetIfPrimaryWindowVisibleAsync()) //if the primary window isn't visible
                 {
-                    if (!await App.GetIfPrimaryWindowVisibleAsync()) //if the primary window isn't visible
+                    if (e.RadioProgram.Style == Core.Stations.StationProgramStyle.Block)
                     {
-                        if (e.RadioProgram.Style == Core.Stations.StationProgramStyle.Block)
-                        {
-                            NepApp.UI.Notifier.ShowStationBlockProgrammingToastNotification(e.RadioProgram, e.Metadata);
-                        }
-                        else
-                        {
-                            NepApp.UI.Notifier.ShowStationHostedProgrammingToastNotification(e.RadioProgram, e.Metadata);
-                        }
+                        NepApp.UI.Notifier.ShowStationBlockProgrammingToastNotification(e.RadioProgram, e.Metadata);
                     }
                     else
                     {
-                        if (e.RadioProgram.Style == Core.Stations.StationProgramStyle.Block)
-                        {
-                            await NepApp.UI.Overlay.ShowSnackBarMessageAsync("Tuning into " + e.RadioProgram.Name + " on " + e.Station);
-                        }
-                        else
-                        {
-                            await NepApp.UI.Overlay.ShowSnackBarMessageAsync("Tuning into " + e.RadioProgram.Name + " by " + e.RadioProgram.Host);
-                        }
+                        NepApp.UI.Notifier.ShowStationHostedProgrammingToastNotification(e.RadioProgram, e.Metadata);
                     }
                 }
-
-                if (e.Metadata != null)
-                    NepApp.UI.Notifier.UpdateLiveTile(new ExtendedSongMetadata(e.Metadata));
+                else
+                {
+                    if (e.RadioProgram.Style == Core.Stations.StationProgramStyle.Block)
+                    {
+                        await NepApp.UI.Overlay.ShowSnackBarMessageAsync("Tuning into " + e.RadioProgram.Name + " on " + e.Station);
+                    }
+                    else
+                    {
+                        await NepApp.UI.Overlay.ShowSnackBarMessageAsync("Tuning into " + e.RadioProgram.Name + " by " + e.RadioProgram.Host);
+                    }
+                }
             }
+
+            if (e.Metadata != null)
+                NepApp.UI.Notifier.UpdateLiveTile(new ExtendedSongMetadata(e.Metadata));
         }
 
         private async void SongManager_SongChanged(object sender, Media.Songs.NepAppSongChangedEventArgs e)
         {
-            if (App.GetDevicePlatform() != Crystal3.Core.Platform.IoT)
+            if (!await App.GetIfPrimaryWindowVisibleAsync()) //if the primary window isn't visible
             {
-                if (!await App.GetIfPrimaryWindowVisibleAsync()) //if the primary window isn't visible
-                {
-                    if ((bool)NepApp.Settings.GetSetting(AppSettings.ShowSongNotifications))
-                        NepApp.UI.Notifier.ShowSongToastNotification((ExtendedSongMetadata)e.Metadata);
-                }
-
-                NepApp.UI.Notifier.UpdateLiveTile((ExtendedSongMetadata)e.Metadata);
+                if ((bool)NepApp.Settings.GetSetting(AppSettings.ShowSongNotifications))
+                    NepApp.UI.Notifier.ShowSongToastNotification((ExtendedSongMetadata)e.Metadata);
             }
+
+            NepApp.UI.Notifier.UpdateLiveTile((ExtendedSongMetadata)e.Metadata);
         }
 
         private void SongManager_PreSongChanged(object sender, Media.Songs.NepAppSongChangedEventArgs e)
@@ -244,21 +185,18 @@ namespace Neptunium.ViewModel
         {
             base.OnNavigatedTo(sender, e);
 
-            if (App.GetDevicePlatform() != Crystal3.Core.Platform.IoT)
-            {
-                var stationsPage = NepApp.UI.NavigationItems.FirstOrDefault(X => X.NavigationViewModelType == typeof(StationsPageViewModel));
-                if (stationsPage == null) throw new Exception("Stations page not found.");
-                NepApp.UI.NavigateToItem(stationsPage);
+            var stationsPage = NepApp.UI.NavigationItems.FirstOrDefault(X => X.NavigationViewModelType == typeof(StationsPageViewModel));
+            if (stationsPage == null) throw new Exception("Stations page not found.");
+            NepApp.UI.NavigateToItem(stationsPage);
 
-                RaisePropertyChanged(nameof(ResumePlaybackCommand));
+            RaisePropertyChanged(nameof(ResumePlaybackCommand));
 
 
-                //CheckForReverseHandoffOpportunitiesIfSupported();
+            //CheckForReverseHandoffOpportunitiesIfSupported();
 
 #if RELEASE
             await CheckForUpdatesAsync();
 #endif
-            }
         }
 
         private static async Task CheckForUpdatesAsync()
